# Move read flow statistics to cursor