<!--
Thank you for contributing to TiKV! Please read TiKV's [CONTRIBUTING](https://github.com/pingcap/tikv/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

## What have you changed? (mandatory)

Previously our flow statistics is not very accurate. For example, when doing `near_seek`, the read during each `next()` attempt is not counted in the statistics. According to PD developers, these read flows should be counted however. Also, previously we self-implemented some read statistics in raw_scan and raw_batch_scan.

This PR moves the read flow statistics logic from mvcc layer into `Cursor` layer, the same as CFStatistics. In detail, when `key()` or `value()` is called, its read flow is counted. Subsequent calls will not be counted unless the cursor is moved (e.g. by `seek()`, `next()`, etc).

This change is extracted from the [Mvcc Refinement PR](https://github.com/pingcap/tikv/pull/3344).

## What are the type of the changes? (mandatory)

- Improvement (non-breaking change which is an improvement to an existing feature)

## How has this PR been tested? (mandatory)

CI

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)

No.

## Does this PR affect tidb-ansible update? (mandatory)

No.

## Refer to a related PR or issue link (optional)

## Benchmark result if necessary (optional)

There is no notable performance difference after applying this PR's change:

master:

```
test storage::scan::bench_seek_ts                                    ... bench:      85,256 ns/iter (+/- 39,536)
test storage::scan::bench_seek_ts                                    ... bench:      82,645 ns/iter (+/- 13,326)
test storage::scan::bench_seek_ts                                    ... bench:      85,052 ns/iter (+/- 26,659)
test storage::scan::bench_seek_ts                                    ... bench:      87,078 ns/iter (+/- 10,894)
```

this PR:

```
test storage::scan::bench_seek_ts                                    ... bench:      84,566 ns/iter (+/- 17,616)
test storage::scan::bench_seek_ts                                    ... bench:      82,349 ns/iter (+/- 14,759)
test storage::scan::bench_seek_ts                                    ... bench:      82,558 ns/iter (+/- 14,742)
test storage::scan::bench_seek_ts                                    ... bench:      88,204 ns/iter (+/- 11,680)
test storage::scan::bench_seek_ts                                    ... bench:      82,717 ns/iter (+/- 12,728)
```

## Add a few positive/negative examples (optional)