@siddontang 2018-08-07
can you refer the statistics in the Cursor directly, it is weird to pass stats to key, value function every where.