@breeswish 2018-08-07
@siddontang We have already have the statistics *almost* everywhere in Cursor, for example, `seek()`, `seek_for_prev()`, `next()`, `prev()`, etc. 🤣 This PR just makes it no survivors. I have an idea for Cursor Statistics refinement after Mvcc refinement, by using Thread Local Storages to avoid passing variables everywhere, which should greatly help ease this situation. I will propose a standalone RFC for that change. However for this PR, let's accept it for now, as a patch solution before total refinement.